Added an Upgrade class
This class makes it easy to perform upgrades. Define the proper output
method (serial or socket) and provide some parameters and the Upgrade
class will perform the complete upgrade.

The defaults are for P4, I2, Mavic & Spark but with some small parameter
tweaks the Mavic RC, Spark RC & Goggles are supported as well. It
should be able to deal with all devices as long as the correct parameters
are used. The main function at the bottom of the file contains examples
of the known devices and their parameters.