DUML: retrying connections, blocking send cmd & debug flag
Connections (Serial & Socket) will now retry for 20 seconds to restore
the connection if it got interrupted (e.g. the device is rebooting).
When upgrading a device, this can happen multiple times and we need to
reconnect to keep track of progress.

A send() method has been added. If the msg is a request that requires an
ACK, the call will block until either the reply to the message is
received or a timeout expires. There's also a register_handler() method
that can be used to register callbacks for requests sent by the device
(e.g. upgrade progress status).

A debug flag has been added to suppress or enable debug output. By
default it's disabled but it can be enabled by setting 'debug: true'
when constructing a DUML object. Beware that messages outside of set 0
are still not printed due to flood (e.g. sticks position on the remote).