DUML: parse the reply of cmd_upgrade_data
This reply can either be 3 bytes long or 263 (no other size seen for
now). A reply of 3 bytes means that it contains the max block size to
use for transfering the upgrade data over duml. The other one contains
the IP address, port number and target filename to use for an upgrade
over FTP.