Upgrade: use new features provided by the updated DUML.rb
Every command will be executed one by one and results of the commands
can be used. As a consequence of this, it's no longer required to
indicate if the upgrade has to be done over ftp or not since the device
will tell us that.

A callback is registered to report the upgrade progress.

The .cfg.sig file is retrieved to parse out the formal version string
(e.g. "1.3.700").

The debug flag has been added and will be passed on from Upgrade to
DUML.

The sleep(120) has been removed as the Upgrade.rb now blocks during the
upgrade and continues when the upgrade is finished.