Upgrade: added a function to get upgrade file info
upgrade_file_info(path) will parse the file, extracts the .cfg.sig file
out of it and gets the device_id and formal_version of the given
dji_system.bin file.

This information can later be used to make sure that you're not trying
to upgrade a Mavic with P4 firmware for example.