FC: read parameter definitions and cache locally
When FlightController talks to the FC, it will request the version
number and check if there's a local json file with parameters for this
version. When it's missing, a new cache will be made by requesting all
definitions from the FC. The next time FlightController talks to this
FC, it will already know all the parameters.